Add files via upload
I've updated the fig file to include the changes and look / feel of the first part of the GUI that we discussed at the last meeting.  We still need to add the rest of the features that are needed for a new participant, but it may be worth merging this with whatever screen will handle updating the participant's profile with new data.